Added 2023 day 7 part 1.

+++ src/integrationTest/java/com/cpt/adventofcode/solution/year2023/day7/Day7Part2Solution2023IntegrationTest.java
@@ -5,13 +5,13 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 @DisplayName("Day 7 Part 2")
-class Day7Part2Solution2023IntegrationTest extends BaseTest<Integer> {
+class Day7Part2Solution2023IntegrationTest extends BaseTest<Long> {
 
     @Test
     @DisplayName("Given test input: should return 0")
     void givenTestInput_ShouldReturn0() {
         Day7Part2Solution2023 solution = new Day7Part2Solution2023();
-        assertSolutionIsCorrect(solution, 0);
+        assertSolutionIsCorrect(solution, 0L);
     }
 
 }
